+++ extensions/dicom-segmentation/src/getOHIFDicomSegSopClassHandler.js
@@ -28,6 +28,7 @@ export default function getSopClassHandlerModule({ servicesManager }) {
       const {
         SeriesDate,
         SeriesTime,
+        SeriesNumber,
         SeriesDescription,
         FrameOfReferenceUID,
         SOPInstanceUID,
@@ -53,6 +54,7 @@ export default function getSopClassHandlerModule({ servicesManager }) {
         isLoaded: false,
         SeriesDate,
         SeriesTime,
+        SeriesNumber,
         SeriesDescription,
         SeriesNumber,
         metadata,